requirements/base.txt
#Forms
django-braces

#Models
django-model-utils

#For the persistance stores
psycopg2

#Cumsom requirements for this project
djangorestframework
markdown
django-filter
